+++ packages/react-theme-default/src/components/SCDefaultHeaderContent.tsx
@@ -1,0 +1,10 @@
+const Component = {
+  styleOverrides: {
+    root: () => ({
+      display: 'flex',
+      alignItems: 'center'
+    })
+  }
+};
+
+export default Component;