+++ packages/react-ui/CHANGELOG.md
@@ -3,6 +3,79 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.52](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.7.52-alpha.1...@selfcommunity/react-ui@0.7.52) (2024-07-25)
+
+**Note:** Version bump only for package @selfcommunity/react-ui
+
+
+
+
+
+## [0.7.52-alpha.1](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.7.52-alpha.0...@selfcommunity/react-ui@0.7.52-alpha.1) (2024-07-25)
+
+**Note:** Version bump only for package @selfcommunity/react-ui
+
+
+
+
+
+## [0.7.52-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.7.51...@selfcommunity/react-ui@0.7.52-alpha.0) (2024-07-25)
+
+**Note:** Version bump only for package @selfcommunity/react-ui
+
+
+
+
+
+## [0.7.51](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.7.51-alpha.6...@selfcommunity/react-ui@0.7.51) (2024-07-25)
+
+**Note:** Version bump only for package @selfcommunity/react-ui
+
+
+
+
+
+## [0.7.51-alpha.6](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.7.51-alpha.5...@selfcommunity/react-ui@0.7.51-alpha.6) (2024-07-25)
+
+**Note:** Version bump only for package @selfcommunity/react-ui
+
+
+
+
+
+## [0.7.51-alpha.5](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.7.51-alpha.4...@selfcommunity/react-ui@0.7.51-alpha.5) (2024-07-25)
+
+
+### Bug Fixes
+
+* DefaultDrawerContent, DefaultHeaderContent ([94b0510](https://github.com/selfcommunity/community-js/commit/94b05108ea01748fa0ca76dcba678c6d82af4b83))
+
+
+
+
+
+## [0.7.51-alpha.4](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.7.51-alpha.3...@selfcommunity/react-ui@0.7.51-alpha.4) (2024-07-25)
+
+
+### Bug Fixes
+
+* DefaultDrawerContent, DefaultHeaderContent ([0e7c6e1](https://github.com/selfcommunity/community-js/commit/0e7c6e15330240cf31444e246f90b51628a5808d))
+
+
+
+
+
+## [0.7.51-alpha.3](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.7.51-alpha.2...@selfcommunity/react-ui@0.7.51-alpha.3) (2024-07-19)
+
+
+### Bug Fixes
+
+* platform widget with tutorial steps ([a3ff798](https://github.com/selfcommunity/community-js/commit/a3ff79813bb0bfa1a04e9324ff33fc04c0c27f43))
+
+
+
+
+
 ## [0.7.51-alpha.2](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-ui@0.7.51-alpha.1...@selfcommunity/react-ui@0.7.51-alpha.2) (2024-07-19)